correccion de PR finalizada
