+++ src/features/profile/edit-profile/edit-profile-modal.tsx
@@ -2,16 +2,17 @@ import { FormEvent, useState, useCallback, useEffect } from "react";
 import { Box, Button, TextField, Modal, IconButton } from "@mui/material";
 import EditIcon from "@mui/icons-material/Edit";
 import useEditProfileFormState from "./use-edit-profile-form-state";
-import { fetchOneProfile, updateProfile } from "../services/profile.service";
 import { useAuthState } from "hooks/use-auth-state";
 import { modalStyle } from "./styles/styles";
+import { updateProfile } from "../services/profile/update-profile.service";
+import useGetProfileInfo from "./use-get-profile-info";
 
 const EditProfileModal = () => {
   const [isOpen, setIsOpen] = useState(false);
   const openModal = () => setIsOpen(true);
   const closeModal = () => setIsOpen(false);
 
-  const { token, user } = useAuthState();
+  const { token } = useAuthState();
 
   const {
     name,
@@ -22,17 +23,7 @@ const EditProfileModal = () => {
     onChangeDescription,
   } = useEditProfileFormState();
 
-  const getUserInfo = useCallback(async () => {
-    try {
-      if (token && user) {
-        const data = await fetchOneProfile(token, user.id);
-        setName(data.user.name);
-        setDescription(data.description);
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  }, [setDescription, setName, token, user]);
+  const { getUserInfo } = useGetProfileInfo({ setName, setDescription });
 
   useEffect(() => {
     getUserInfo();
@@ -42,16 +33,15 @@ const EditProfileModal = () => {
     async (event: FormEvent<HTMLFormElement>) => {
       event.preventDefault();
       try {
-        if (token && user) {
-          const data = await updateProfile(token, { name, description });
-          setIsOpen(false);
-          return data;
-        }
+        if (!token || !name || !description) return;
+        const data = await updateProfile(token, { name, description });
+        setIsOpen(false);
+        return data;
       } catch (error) {
         console.log(error);
       }
     },
-    [description, name, token, user]
+    [description, name, token]
   );
 
   return (
