+++ src/features/profile/services/profile/update-profile.service.ts
@@ -1,0 +1,20 @@
+import { PROFILES_URL } from "./config/config";
+import { MessageResponse, UpdateProfileBody } from "./config/config";
+import axios from "axios";
+import { BackendError } from "app/exceptions";
+
+export async function updateProfile(
+  token: string,
+  body: UpdateProfileBody
+): Promise<MessageResponse> {
+  try {
+    const response = await axios.patch(`${PROFILES_URL}/my-profile`, body, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    return response.data;
+  } catch (error) {
+    throw new BackendError(error);
+  }
+}
