+++ src/features/profile/services/skills/config/config.ts
@@ -1,0 +1,2 @@
+import { BACKEND_V1_URL } from "app/config";
+export const SKILLS_URL = `${BACKEND_V1_URL}/skills`;
