+++ src/features/profile/services/skills/get-all-skills.service.ts
@@ -1,13 +1,11 @@
-import { BACKEND_V1_URL } from "app/config";
-import { BackendError } from "app/exceptions";
-import axios from "axios";
+import { SKILLS_URL } from "./config/config";
 import { Skill } from "core/profiles/types";
-
-const URL = `${BACKEND_V1_URL}/skills`;
+import axios from "axios";
+import { BackendError } from "app/exceptions";
 
 export async function getAllSkills(token: string): Promise<Skill[]> {
   try {
-    const response = await axios.get(URL, {
+    const response = await axios.get(SKILLS_URL, {
       headers: {
         Authorization: `Bearer ${token}`,
       },
